Updated prometheus monitoring for transactions.
As I have used the helm chart for the prometheus community stack,
which in turn uses the prometheus monitor, I had to deploy a
ServiceMonitor for that operator to be able to idenitfy my service
and grab metrics.  I also had to change the labels so they'd
match up.

This latter part, the lable matching, is still a bit confusing
to me.  But, this works.

See https://github.com/prometheus-operator/prometheus-operator/blob/master/Documentation/design.md#servicemonitor
for not enough detail about these service monitors.